test: adapt tests to work with one tilequeue per layer
Introduced an OL TileQueue mock since we cannot be creating the tilequeue
ourselves
